ejb refactor (#2528)
* ejb refactor

* Remove unused module
