+++ ejb/ejb-remote/src/main/java/com/baeldung/ejb/wildfly/TextProcessorRemote.java
@@ -1,4 +1,4 @@
-package com.baeldung.ejbmodule;
+package com.baeldung.ejb.wildfly;
 
 import javax.ejb.Remote;
