--- ejb/ejbclient/src/main/resources/jboss-ejb-client.properties
@@ -1,8 +1,0 @@
-endpoint.name=client-endpoint
-remote.connectionprovider.create.options.org.xnio.Options.SSL_ENABLED=false
-remote.connections=default
-remote.connection.default.host=127.0.0.1
-remote.connection.default.port=8080
-remote.connection.default.connect.options.org.xnio.Options.SASL_POLICY_NOANONYMOUS=false
-remote.connection.default.username=myusername
-remote.connection.default.password=mypassword
